+++ ejb/src/main/java/net/sf/ahtutils/model/interfaces/status/UtilsStatus.java
@@ -4,8 +4,8 @@ import java.util.Map;
 
 public interface UtilsStatus<T extends UtilsLang>
 {	
-	public void setId(int id);
-	public int getId();
+	public void setId(long id);
+	public long getId();
 	
 	public Map<String, T> getName();
 	public void setName(Map<String, T> name);
